csec_hub/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
bleach==5.0.0
Django==4.0.4
django-filter==21.1
django-summernote==0.8.20.0
django-taggit==3.0.0
Pillow==9.1.0
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
webencodings==0.5.1
